add schedule2 page
